refactor: Refactor Workbench
