+++ components/Workbench/Job.tsx
@@ -1,20 +1,37 @@
 import { FC } from 'react';
 
-import { VStack, Heading } from '@chakra-ui/react';
+import { Flex, VStack, Heading, Divider, Button } from '@chakra-ui/react';
+
+import { FaCloudDownloadAlt } from 'react-icons/fa';
 
 type JobProps = {
-    table: string;
+    job: string;
 };
 
-const Job: FC = () => {
+const Job: FC<JobProps> = ({ job }) => {
     return (
         <VStack
-            maxH="full"
+            h="50%"
             w="full"
+            flex="1"
             alignItems="stretch"
             borderWidth={1}
             p={4}
-        ></VStack>
+            spacing={4}
+        >
+            <Heading size="md">{job || 'Job'}</Heading>
+            <Divider />
+            <Flex w="full" justifyContent="flex-end">
+                <Button
+                    rightIcon={<FaCloudDownloadAlt fill="white" />}
+                    colorScheme="blue"
+                    // isLoading={loading}
+                    // onClick={handleRequest}
+                >
+                    Request
+                </Button>
+            </Flex>
+        </VStack>
     );
 };
 export default Job;
