+++ components/Workbench/Request.tsx
@@ -1,4 +1,4 @@
-import { FC, useState } from 'react';
+import { Dispatch, FC, SetStateAction, useState } from 'react';
 
 import {
     VStack,
@@ -7,75 +7,31 @@ import {
     Divider,
     Heading,
     Button,
-    Popover,
-    PopoverTrigger,
-    PopoverContent,
-    PopoverBody,
-    PopoverArrow,
-    Icon,
 } from '@chakra-ui/react';
-import {
-    DayPicker,
-    DateRange,
-    SelectRangeEventHandler,
-} from 'react-day-picker';
-import 'react-day-picker/dist/style.css';
-import { formatISO } from 'date-fns';
+import { DateRange } from 'react-day-picker';
+import { format } from 'date-fns';
 
-import { FaRegCalendarAlt, FaCloudUploadAlt } from 'react-icons/fa';
+import { FaCloudUploadAlt } from 'react-icons/fa';
+
+import PopoverDatePicker, { defaultDateRange } from './DatePicker';
 
 type RequestProps = {
     table: string;
+    setJob: Dispatch<SetStateAction<string>>;
 };
 
-type DatePickerProps = {
-    range?: DateRange;
-    setRange: SelectRangeEventHandler;
-};
+const Request: FC<RequestProps> = ({ table, setJob }) => {
+    const [range, setRange] = useState<DateRange | undefined>(defaultDateRange);
 
-const PopoverDatePicker: FC<DatePickerProps> = ({ range, setRange }) => {
-    const { from, to } = range || { from: new Date(), to: new Date() };
+    const [loading, setLoading] = useState(false);
 
-    return (
-        <Popover>
-            <PopoverTrigger>
-                <Button w="full" leftIcon={<Icon as={FaRegCalendarAlt} />}>
-                    {[from, to]
-                        .map(
-                            (value) =>
-                                value &&
-                                formatISO(value, { representation: 'date' })
-                        )
-                        .join(' - ')}
-                </Button>
-            </PopoverTrigger>
-            <PopoverContent>
-                <PopoverArrow />
-                <PopoverBody>
-                    <DayPicker
-                        mode="range"
-                        fromYear={2018}
-                        toYear={2023}
-                        captionLayout="dropdown"
-                        selected={range}
-                        onSelect={setRange}
-                        modifiersStyles={{
-                            selected: {
-                                backgroundColor: '#3182CE',
-                            },
-                        }}
-                    />
-                </PopoverBody>
-            </PopoverContent>
-        </Popover>
-    );
-};
-
-const Request: FC<RequestProps> = ({ table }) => {
-    const [range, setRange] = useState<DateRange | undefined>({
-        from: new Date(),
-        to: new Date(),
-    });
+    const handleRequest = () => {
+        setLoading(true);
+        setTimeout(() => {
+            setJob(format(new Date(), 'yyyy-MM-ddTHH:mm:ss.SSS'));
+            setLoading(false);
+        }, 1000);
+    };
 
     return (
         <VStack
@@ -85,17 +41,20 @@ const Request: FC<RequestProps> = ({ table }) => {
             alignItems="stretch"
             borderWidth={1}
             p={4}
+            spacing={4}
         >
-            <Heading pb={2} size="md">{table || 'Table'}</Heading>
+            <Heading size="md">{table || 'Table'}</Heading>
             <Divider />
             <HStack justifyContent="stretch">
-                <PopoverDatePicker
-                    range={range}
-                    setRange={setRange}
-                />
+                <PopoverDatePicker range={range} setRange={setRange} />
             </HStack>
-            <Flex w="full" pt={2} justifyContent="flex-end">
-                <Button rightIcon={<FaCloudUploadAlt fill="white" />} colorScheme="blue">
+            <Flex w="full" justifyContent="flex-end">
+                <Button
+                    rightIcon={<FaCloudUploadAlt fill="white" />}
+                    colorScheme="blue"
+                    isLoading={loading}
+                    onClick={handleRequest}
+                >
                     Request
                 </Button>
             </Flex>
